requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
gunicorn==21.2.0
```